2007-01-20 Kimmo Varis <kimmov@winmerge.org> 	* scew/error.c: Fix compile with VS2005

+++ Externals/scew/scew/error.c
@@ -62,7 +62,7 @@ scew_error_string(scew_error code)
         // This is not thread safe. Even though, no one should get in
         // here.
         static XML_Char unk_message[35];
-        scew_sprintf(unk_message, _XT("Unknown error code (%d)"), code);
+        scew_sprintf(unk_message, _XT("Unknown error code (%d)"), (int) code);
         return unk_message;
     }
     else
